Fix the .update_observations() again
The .update_observations() function was comparing the wrong (x, y) positions when calculating the rover observation portion of the state. The signature of .add_to_sensor() specifically uses the *other* (x, y), not the self.

This needed to be changed to actually properly measure the observations of the other rovers.