Implemented Multi-policy meta learning
Creates 3 policies that simply go straight toward a POI of type X.
This creates a simpler problem for the meta learner to solve.
The sub-policies can be learned; I didn't want to fiddle with training them right now, so they are written out.

Meta-policy learner quickly gets to all 3 POI types, and recieves the full reward.